We write the test function for the Hello function.
Hello function prints out a greeting for the name of the person passed,
and prints out a "Hello, World!", when no name is passed. The Hello
function has a dependency of printing, so we solve this problem by
Dependency Injection. We don't care about where the printing happens so
in our test function we pass an interface to the Hello function, which
will store whatever printing happens.